TODO.md: that's no race condition
All indicates that the process is being killed due to being OOM (udge-sandbox)

The change in setting from 512M to 1G should have fixed this.

A good test to discover this was running just "make test-parallel -j7" with
bary any application open, then changing to run it with a browser with dozens
of tabs open.  When doing the latter the error happened almost always.

In practice, when running everything sandboxed at a different user this problem
should not appear.

This is a 95% guess.  I'll keep an eye open if the problem happens again.