improve detection of output limit exceeded
When a submission was made to a problem with at least 6 runs sometime Udge
would fail to detect it and would report "unknown error".

This happened because each of the 6 runs would write around 2MiB totaling
12MiB.  This would be fine as it is with a disk limit of 12MiB, however, when
the output is copied before checking, it would go over this limit thus yielding
the "unknown error".

With this patch, Udge checks for output limit exceeded errors earlier
and cleans up any created files so that the reason for failure is reported
correctly as "output limit exceeded".