drop Erlang support
Erlang support in Udge turned out to be too hard too maintain.
Several hacks had to be included in the code to make it work.
(See the commit diff for details.)

In cscx.org nobody ever used Erlang in the (almost) couple of years it was
deployed between 2021-12 and 2023-08...

So Erlang support is dropped for the time being.  :-(