Add regression test (.hs misidentified as .ini)
Consider the following example:

	$ cat | sed 's/$/^M/' >file.hs

	fun :: [a] -> [a]
	^D

	$ file file.hs
	file.hs: Generic INItialization configuration [a]

Programs with DOS line breaks that start on a blank line and contain
list-to-list functions are misidentified by the file program as ".ini"
configuration files.

This ultimately causes Udge to incorrectly report "compile error"
to solutions that are ultimately correct.  (Please see the diff for more
details and a full example.)

Many thanks to Elias Dovkrans for finding and reporting this issue.

This commit is not a fix, but merely an inclusion of a regression test.
The fix will be performed in a future commit.