Fix incorrect "compile error" for some programs
Some Haskell programs with CRLF (DOS) line terminators
are misidentified by file as ".ini" config files.

This commits makes it so that the relevant call to file has the argument -esoft
which deactivates consulting of magic files.  This causes most computer
programs to be labeled as "text/plain".  In the context of Udge, this is
acceptable.

The regression test added on 4e287b205ce640958fb325bc2420b0245f5d8f6a was
updated accordingly.

Thanks to Elias Dovkrans for making me aware of the original issue.