disable UsePerfData when compiling java
There was an issue with running udge-judge in parallel for java submissions.
This would cause Makefiles with lots of .clitests involving java programs
to fail randomly.

Consider the following script:

	#!/bin/bash
    udge-judge hello-world examples/hello-world/hello-world.java -v &
    udge-judge hello-world examples/hello-world/hello-world.java -v &
    udge-judge hello-world examples/hello-world/hello-world.java -v &
    udge-judge hello-world examples/hello-world/hello-world.java -v &
    wait

Only one of the above would get a correct 1/1 score,
whereas the others would all get 0/1 incorrect output.

The issue is revealed in the diff output:

	$ ./script
	...
	--- /var/lib/udge/problem/hello-world/out 2022-01-12 20:18:58.609172856 +0100
	+++ /tmp/judge.GNFZhJZxh5/out/out 2023-12-08 18:07:31.173061737 +0100
	@@ -1 +1,2 @@
	+[0.001s][warning][perf,memops] Cannot use file /tmp/hsperfdata_root/3 because it is locked by another process (errno = 11)
	 Hello, World!

Java programs share the /tmp/hsperfdata_root folder!
If they cannot open it due to a lock,
stdout is polluted with the above warning message.

We can turn this off by passing "-XX:-UsePerfData" to java,
which is what this patch/commit does.

If a similar problem appears again, the above test script can be useful,
along with the following Makefile target:

	# Java sometimes has issues with parallel submissions.
	# This target can be used to test this semi consistently.
	#
	#     $ make -j3 test-java-parallel
	test-java-parallel: \
	  add-java.clitest \
	  rectangle-java.clitest \
	  hello-world-java.clitest

Further information can be found on the following links:

* https://stackoverflow.com/questions/76327/how-can-i-prevent-java-from-creating-hsperfdata-files
* https://docs.oracle.com/javase/8/docs/technotes/tools/unix/java.html
* https://bugs.java.com/bugdatabase/view_bug?bug_id=5012932